requirements.txt
Jinja2==2.9
Flask==0.12
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-RESTful==0.3.5
flask-cors==3.0.2
flask-debugtoolbar==0.10.1
flask-migrate==2.0.3

SQLAlchemy==1.1.4
psycopg2==2.6.2

itsdangerous==0.24

gunicorn==19.6.0
pytz==2016.10

pep8==1.7.0

PyYaml==3.12
